kubectl plugin: change token cache behaviour
Now we will pass a directory path instead of the actual file to NewToken(),
we also default to the same path as the environment variable KUBECONFIG
if it's defined.

We also remove a where we check the user home directory in a better way
to support more than just Linux.